+++ CHANGELOG.md
@@ -44,6 +44,7 @@
 
 ### Chore
 
+- [#9588](https://github.com/blockscout/blockscout/pull/9588) - Enable URL-encoded chars support in DB username
 - [#9571](https://github.com/blockscout/blockscout/pull/9571) - Support Optimism Ecotone upgrade by Indexer.Fetcher.Optimism.TxnBatch module
 - [#9562](https://github.com/blockscout/blockscout/pull/9562) - Add cancun evm version
 - [#9506](https://github.com/blockscout/blockscout/pull/9506) - API v1 bridgedtokenlist endpoint